+++ library/docs/changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Upgrade to DataStore `1.0.0-alpha04`.
+
 ## [0.9.1] - 2020-11-12
 
 ### Added