+++ library/docs/changelog.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BlockingLaboratory` class that can read and write feature flags via blocking API.
 - `blocking()` function to `Laboratory` class that is an entry point to the blocking API.
 
+### Changed
+- Upgrade to DataStore `1.0.0-alpha04`.
+
 ### Deprecated
 - All blocking functions on the `Laboratory` class. `BlockingLaboratory` available via `blocking()` function should be used instead.